requirements.txt
numpy>=1.20.0
open3d>=0.15.0
scipy>=1.7.0
opencv-python>=4.5.0
PyYAML>=5.4.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
tqdm>=4.62.0